+++ R/arc-read.R
@@ -13,10 +13,13 @@
 #'   the feature. If `col_names` is a character vector with the same length as
 #'   the number of columns in the layer, the default names are replaced with the
 #'   new names. If `col_names` has one fewer name than the default column names,
-#'   the existing sf column name is retained. If `col_names` is the string
-#'   `"alias"`, names are set to match the alias names for the layer, if available.
+#'   the existing sf column name is retained.
 #' @param col_select Default `NULL`. A character vector of the field names to be
 #'   returned. By default, all fields are returned.
+#' @param alias Default "drop". Supported options: `c("drop", "label",
+#'   "replace")`. If "drop", field alias values are ignored. If "label", field
+#'   alias values are assigned as a label attribute for each field. If replace,
+#'   field alias values are used as the column names.
 #' @param n_max Defaults to 10000 or an option set with
 #'   `options("arcgislayers.n_max" = <max records>)`. Maximum number of records
 #'   to return.
@@ -44,7 +47,7 @@
 #'   )
 #'
 #'  # use field aliases as column names
-#'  arc_read(furl, col_names = "alias")
+#'  arc_read(furl, alias = "replace")
 #'
 #'  # read an ImageServer directly
 #'  img_url <- "https://landsat2.arcgis.com/arcgis/rest/services/Landsat/MS/ImageServer"
@@ -68,6 +71,7 @@ arc_read <- function(
     crs = NULL,
     ...,
     fields = NULL,
+    alias = c("drop", "label", "replace"),
     token = arc_token()
 ) {
   service <- arc_open(url = url, token = token)
@@ -106,72 +110,136 @@ arc_read <- function(
     ...
   )
 
-  set_layer_names(
+  set_layer_col_names(
     layer,
+    service = service,
     col_names = col_names,
     name_repair = name_repair,
-    alias = service[["fields"]][["alias"]]
+    alias = alias
   )
 }
 
 #' Set names for layer or table
 #'
 #' @noRd
-set_layer_names <- function(
+set_layer_col_names <- function(
     x,
+    service = NULL,
     col_names = NULL,
     name_repair = NULL,
-    alias = NULL,
+    alias = c("drop", "label", "replace"),
     call = rlang::caller_env()
 ) {
 
   # Use existing names by default
-  layer_nm <- names(x)
-  nm <- layer_nm
+  field_nm <- names(x)
+  replace_nm <- field_nm
   sf_column_nm <- attr(x, "sf_column")
 
-  if (is.character(col_names)) {
-    # Assign alias values as name if col_names = "alias"
-    # FIXME: What if the data has a column named alias?
-    if (identical(col_names, "alias")) {
-      col_names <- alias
-    }
+  alias <- rlang::arg_match(alias, error_call = call)
 
-    nm <- col_names
+  if (alias == "label") {
+    # get fields from service object
+    fields <- service[["fields"]]
+
+    # filter alias values to selected fields
+    fields <- fields[fields[["name"]] %in% field_nm, ]
+
+    # get alias values
+    alias_val <- fields[["alias"]]
   }
 
-  nm_len <- length(nm)
+  if (is.character(col_names)) {
+    if (identical(col_names, "alias")) {
+      # Assign alias values as name if col_names = "alias"
+      col_names <- alias_val
+      lifecycle::signal_stage(
+        "superseded",
+        what = "arc_read(..., field = 'alias')",
+        with = "arc_read(..., alias = 'replace')",
+      )
+    }
+
+    if (alias == "replace") {
+      # NOTE: alias values may not be valid names
+      col_names <- alias_val
+    }
+
+    replace_nm <- col_names
+  }
+
+  replace_nm_len <- length(replace_nm)
 
   if (rlang::is_false(col_names)) {
     # Use X1, X2, etc. as names if col_names is FALSE
-    nm <- paste0("X", seq(to = nm_len))
+    replace_nm <- paste0("X", seq(to = replace_nm_len))
   }
 
   # If x is a sf object and sf column is not in names, check to ensure names
   # work with geometry column
-  if (inherits(x, "sf") && sf_column_nm != nm[[nm_len]]) {
-    layer_nm_len <- length(layer_nm)
-    if (length(nm) == layer_nm_len) {
+  if (inherits(x, "sf") && sf_column_nm != replace_nm[[replace_nm_len]]) {
+    field_nm_len <- length(field_nm)
+    if (length(replace_nm) == field_nm_len) {
       # If same number of names as layer columns, use last name for geometry
-      x <- sf::st_set_geometry(x, nm[[length(layer_nm)]])
-    } else if (nm_len == (layer_nm_len - 1)) {
+      x <- sf::st_set_geometry(x, replace_nm[[length(field_nm)]])
+    } else if (replace_nm_len == (field_nm_len - 1)) {
       # If same number of names as layer columns, use existing geometry name
-      nm <- c(nm, sf_column_nm)
+      replace_nm <- c(replace_nm, sf_column_nm)
     }
   }
 
   if (!is.null(name_repair)) {
     rlang::check_installed("vctrs", call = call)
     nm <- vctrs::vec_as_names(
-      names = nm,
+      names = replace_nm,
       repair = name_repair,
       repair_arg = "name_repair",
       call = call
     )
   }
 
-  rlang::set_names(
+  layer <- rlang::set_names(
     x,
-    nm = nm
+    nm = replace_nm
   )
+
+  if (alias != "label") {
+    return(layer)
+  }
+
+  # Name alias values with layer names
+  alias_val <- rlang::set_names(
+    alias_val,
+    nm = setdiff(replace_nm, attr(layer, "sf_column"))
+  )
+
+  set_layer_col_labels(
+    layer,
+    values = alias_val
+  )
+}
+
+#' @noRd
+set_layer_col_labels <- function(
+    layer,
+    values,
+    safe = TRUE) {
+  if (safe) {
+    nm <- intersect(names(values), names(layer))
+  }
+
+  for (v in nm) {
+    label_attr(layer[[v]]) <- values[[v]]
+  }
+
+  layer
+}
+
+#' Set label attribute
+#' @seealso [labelled::set_label_attribute()]
+#' @source <https://github.com/cran/labelled/blob/master/R/var_label.R>
+#' @noRd
+`label_attr<-` <- function(x, value) {
+  attr(x, "label") <- value
+  x
 }